requirements.txt
matplotlib
aiohttp
discord
python-dotenv
uwuify
requests
Red-DiscordBot